chore: renamed file to fallback-values.ts
